requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asn1crypto==1.5.1
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dataclasses-json==0.6.7
distro==1.9.0
environs==14.1.1
et_xmlfile==2.0.0
feedparser==6.0.11
filelock==3.18.0
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
groq==0.20.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.46
langchain-groq==0.3.0
langchain-openai==0.3.9
langchain-text-splitters==0.3.7
langgraph==0.3.18
langgraph-checkpoint==2.0.21
langgraph-prebuilt==0.1.3
langgraph-sdk==0.1.58
langsmith==0.3.18
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
msgpack==1.1.0
multidict==6.2.0
mypy-extensions==1.0.0
narwhals==1.31.0
numpy==2.2.4
openai==1.67.0
openpyxl==3.1.5
orjson==3.10.15
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.7
propcache==0.3.0
protobuf==6.30.1
pyarrow==19.0.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.32.0
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.23.1
selenium==4.29.0
sgmllib3k==1.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowflake-connector-python==3.14.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.39
streamlit==1.43.2
tenacity==9.0.0
tiktoken==0.9.0
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.18.3
zstandard==0.23.0
streamlit-aggrid
